initial commit
This commit initalizes the github repository with a springboot starter,
a mute entry model, a mute entry repository and a controller for
retrieving (but not yet creating) mute entries.

For some reason the mute entry model is called Mute and and not
MuteEntry, even though I keep refering to it as MuteEntry. Should
probably change that.